requirements.txt
numpy==1.21.4
pandas==1.3.4
Pillow==8.4.0
pip==21.2.3
plotly==5.3.1
setuptools==57.4.0
streamlit==1.1.0
wheel==0.37.0
